Automatically deploy to prod via GitHub Actions (#2)
Right now, I'm managing deployments to prod manually. I've been thinking for a while about automating this via CI/CD and GitHub Actions is free so... This PR sets up a simple GitHub Action to deploy the website to prod on pushes to main. To facilitate this:

the CDK code was updated to generate a new IAM User with minimal permissions
upload.sh was updated to default to deploying to devBucket
package-lock.json was committed so setup-node could effectively cache the NPM install step